Improved uninstall routine, tested with WordPress 3.9.2
git-svn-id: https://plugins.svn.wordpress.org/recent-posts-widget-with-thumbnails/trunk@963447 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -21,5 +21,22 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 * @since   2.0
 */
 // remove settings
-delete_option( 'widget_recent-posts-widget-with-thumbnails' );
+if ( is_multisite() ) {
+
+	$sites = wp_get_sites();
+
+	if ( empty ( $sites ) ) return;
+
+	foreach ( $sites as $site ) {
+		// switch to next blog
+		switch_to_blog( $site[ 'blog_id' ] );
+		// remove settings
+		delete_option( 'widget_recent-posts-widget-with-thumbnails' );
+	}
+	// restore the current blog, after calling switch_to_blog()
+	restore_current_blog();
+} else {
+	// remove settings
+	delete_option( 'widget_recent-posts-widget-with-thumbnails' );
+}
